Remove Inloggen route and refactor Home component for user login; add loading state and error handling
